adding hibernation
hibernate after connecting to  aws mqtt broker it uses the shadow to get latest version and state. from there it  will update and then got to sleep for a certain number f tick, will figure out the right calculation to make it sleep for a certain time value instead of number of ticks.

shadow get rejects when the shadow is deleted since it no longer exists, but when updated the version number doesnt reset so if deleted it must be updated without a version number and then it will be able to do get and then update using the correct version